// src/activity.interface.ts
export interface Activity {
    activity:      string;
    availability:  number;
    type:          string;
    participants:  number;
    price:         number;
    accessibility: string;
    duration:      string;
    kidFriendly:   boolean;
    link:          string;
    key:           string;
}